create types.ts for on specific folder and fix import path
