+++ src/API/articles/types.ts
@@ -1,0 +1,6 @@
+export interface ArticleProps {
+    display_id: number;
+    title: string;
+    img: string;
+    description_list: string[];
+}
