+++ src/API/production-by/types.ts
@@ -1,0 +1,15 @@
+export interface PromoProps {
+    id: number;
+    name: string;
+    photo_url: string;
+    promo_special: number;
+}
+
+export interface ProductionProps {
+    id: number;
+    name: string;
+    photo_url: string;
+    description_list: string[];
+    category: string;
+    created_at: string;
+}
