+++ src/API/tag-categories/types.ts
@@ -1,0 +1,6 @@
+export interface CategoryProps {
+    id: number;
+    name: string;
+    description_list: string[];
+    created_at: string;
+}
