+++ src/app/components/hompage/Article_Section/index.tsx
@@ -1,6 +1,6 @@
 import { AccordionExpandDefault } from "@/app/components";
 import { useArticles } from "@/API/hooks/useArticles";
-import { ArticleProps } from "@/API/types/apiTypes";
+import { ArticleProps } from "@/API/articles/types";
 import AccordionSkeleton from "@/components/ui/AccordionExpandDefault/AccordionSkeleton";
 import Footer from "../../layout/Footer";
