+++ src/app/components/hompage/Production_Section/index.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { ProductionCard } from "@/app/components";
 import { GoChevronDown } from "react-icons/go";
 import { useProductions } from "@/API/hooks/useProductions";
-import { ProductionProps } from "@/API/types/apiTypes";
+import { ProductionProps } from "@/API/production-by/types";
 import ProductionCardSkeleton from "@/app/components/hompage/ProductionCard/ProductionCardSkeleton";
 
 const Production = () => {
